debugged issue 2
had to move pieces back to the container with the class ".puzzle-pieces" since that is where you drag from.
used plzPieces since that is where all the imgs are, collected each piece with forEach, added function for the container where I want the pieces to go, and to SNATCH the pieces by using appendChild, when the function changeImageSet happens.